refactor(playground-ui): move assistant-ui to lib/ai-ui
