+++ packages/playground-ui/src/index.ts
@@ -81,10 +81,10 @@ export * from './lib/tanstack-query';
 export * from './domains/memory/hooks';
 export * from './store/playground-store';
 export * from './lib/framework';
-export { MemorySearch } from './components/assistant-ui/memory-search';
+export { MemorySearch } from './lib/ai-ui/memory-search';
 export * from './domains/conversation/index';
 export * from './lib/errors';
-export * from './components/assistant-ui/tools/tool-fallback';
+export * from './lib/ai-ui/tools/tool-fallback';
 export * from './domains/workflows/runs/workflow-run-list';
 export * from './domains/mcps/index';
 export * from './lib/toast';
