+++ packages/playground-ui/src/lib/ai-ui/attachments/attachment-preview-dialog.tsx
