+++ packages/playground-ui/src/lib/ai-ui/attachments/pdfs-adapter.ts
