+++ packages/playground-ui/src/lib/ai-ui/hooks/use-adapters.ts
