+++ packages/playground-ui/src/lib/ai-ui/hooks/use-attachment-src.ts
