+++ packages/playground-ui/src/lib/ai-ui/hooks/use-file-src.ts
