+++ packages/playground-ui/src/lib/ai-ui/hooks/use-has-attachments.ts
