+++ packages/playground-ui/src/lib/ai-ui/messages/error-aware-text.tsx
