+++ packages/playground-ui/src/lib/ai-ui/messages/markdown-text.tsx
@@ -8,7 +8,7 @@ import '@assistant-ui/react-markdown/styles/dot.css';
 import { CheckIcon, CopyIcon } from 'lucide-react';
 import { FC, ImgHTMLAttributes, memo, useEffect, useState } from 'react';
 import remarkGfm from 'remark-gfm';
-import { TooltipIconButton } from '@/components/assistant-ui/tooltip-icon-button';
+import { TooltipIconButton } from '../tooltip-icon-button';
 import { cn } from '@/lib/utils';
 import { coldarkDark } from 'react-syntax-highlighter/dist/cjs/styles/prism';
 import { makePrismAsyncLightSyntaxHighlighter } from '@assistant-ui/react-syntax-highlighter';
