+++ packages/playground-ui/src/lib/ai-ui/messages/tripwire-notice.tsx
