+++ packages/playground-ui/src/lib/ai-ui/tools/badges/agent-badge.tsx
