+++ packages/playground-ui/src/lib/ai-ui/tools/badges/loading-badge.tsx
