+++ packages/playground-ui/src/lib/ai-ui/tools/badges/network-choice-metadata-dialog.tsx
