+++ packages/playground-ui/src/lib/ai-ui/tools/badges/tool-approval-buttons.tsx
