+++ packages/playground-ui/src/lib/ai-ui/tools/badges/tool-badge.tsx
