+++ packages/playground-ui/src/lib/ai-ui/tooltip-icon-button.tsx
